chore(API): adds others tests & refactoring files names [fixes #167195891]
